:new: Port getEslintConfig to babel
